+++ test/com/sun/jna/StructureTest.java
@@ -895,6 +895,12 @@ public class StructureTest extends TestCase {
         assertEquals("Non-volatile field should be written", 1, s.getPointer().getInt(4));
         s.writeField("counter");
         assertEquals("Explicit volatile field write failed", 1, s.getPointer().getInt(0));
+
+        s.equals(s);
+        assertEquals("Structure equals should leave volatile field unchanged", 1, s.getPointer().getInt(0));
+
+        s.hashCode();
+        assertEquals("Structure equals should leave volatile field unchanged", 1, s.getPointer().getInt(0));
     }
 
     public static class StructureWithPointers extends Structure {
@@ -1542,53 +1548,27 @@ public class StructureTest extends TestCase {
             protected List getFieldOrder() {
                 return Arrays.asList(new String[] { "first", "second", "third" });
             }
+            public TestStructure() { }
+            public TestStructure(Pointer p) { super(p); }
         }
         OtherStructure s0 = new OtherStructure();
         TestStructure s1 = new TestStructure();
-        TestStructure s2 = new TestStructure();
-        TestStructure s3 = new TestStructure();
-        int VALUE = 99;
-        s1.first = s2.first = s3.first = VALUE;
+        TestStructure s2 = new TestStructure(s1.getPointer());
+        TestStructure s3 = new TestStructure(s2.getPointer());
+        TestStructure s4 = new TestStructure();
 
         assertFalse("Structures of different classes with same fields are not equal", s1.equals(s0));
         assertFalse("Structures of different classes with same fields are not equal (reflexive)", s0.equals(s1));
 
         assertFalse("Compare to null failed", s1.equals(null));
         assertTrue("Equals is not reflexive", s1.equals(s1));
-        assertTrue("Equals failed on identical structures", s1.equals(s2));
+        assertTrue("Equals failed on structures with same pointer", s1.equals(s2));
         assertTrue("Equals is not symmetric", s2.equals(s1));
         assertTrue("Equals is not transitive", s1.equals(s2) && s2.equals(s3) && s1.equals(s3));
-
-
+        assertFalse("Compare to different structure failed", s1.equals(s4));
     }
 
-    public void testStructureEqualsByValueByReference() {
-        class TestStructure extends Structure {
-            public int first;
-            public int[] second = new int[4];
-            public Pointer[] third = new Pointer[4];
-            protected List getFieldOrder() {
-                return Arrays.asList(new String[] { "first", "second", "third" });
-            }
-        }
-        class ByReference extends TestStructure implements Structure.ByReference { }
-        class ByValue extends TestStructure implements Structure.ByValue { }
-        TestStructure s1 = new TestStructure();
-        TestStructure s2 = new ByReference();
-        TestStructure s3 = new ByValue();
-        int VALUE = 99;
-        s1.first = s2.first = s3.first = VALUE;
-
-        assertTrue("Equals failed on identical ByReference", s1.equals(s2));
-        assertTrue("Equals is not symmetric (ByReference)", s2.equals(s1));
-        assertTrue("Equals failed on identical ByValue", s1.equals(s3));
-        assertTrue("Equals is not symmetric (ByValue)", s3.equals(s1));
-        assertTrue("Equals is not transitive (ByReference/ByValue)", s1.equals(s2) && s2.equals(s3) && s1.equals(s3));
-
-
-    }
-
-    public void testStructureHashCodeMatchesEqualsTrue() {
+    public void testStructureHashCodeMatchesWhenEqual() {
         class TestStructure extends Structure {
             public int first;
             protected List getFieldOrder() {
@@ -1597,26 +1577,15 @@ public class StructureTest extends TestCase {
         }
         TestStructure s1 = new TestStructure();
         TestStructure s2 = new TestStructure();
-        s1.first = s2.first = 0x12345678;
-        assertEquals("hashCode should match when structures equal",
+        assertFalse("hashCode should be different for two different structures", s1.hashCode() == s2.hashCode());
+
+        s2.useMemory(s1.getPointer());
+        assertEquals("hashCode should match when structures have same pointer",
                      s1.hashCode(), s2.hashCode());
-    }
 
-    public void testStructureEqualsIgnoresPadding() {
-        class TestStructure extends Structure {
-            public byte first;
-            public int second;
-            protected List getFieldOrder() {
-                return Arrays.asList(new String[] { "first", "second" });
-            }
-        }
-        TestStructure s1 = new TestStructure();
-        TestStructure s2 = new TestStructure();
-
-        // Make padding bits non-zero
-        s2.getPointer().setInt(0, -1);
-        s2.write();
-        assertTrue("Structure equals should ignore padding", s1.equals(s2));
+        s2.useMemory(s1.getPointer());
+        assertEquals("hashCode should match when structures have same pointer",
+                     s1.hashCode(), s2.hashCode());
     }
 
     public void testRecursiveWrite() {